+++ packages/core/src/storybook/decorators/withLiveEdit/LiveContent/LiveContent.types.ts
@@ -1,0 +1,8 @@
+import { Decorator, StoryContext } from "@storybook/react";
+
+export interface LiveContentProps {
+  code: string;
+  scope: Record<string, unknown>;
+  decorators: Decorator[];
+  context: StoryContext;
+}