+++ packages/core/src/storybook/decorators/withLiveEdit/withLiveEdit.tsx
@@ -1,4 +1,4 @@
-import { Decorator, Parameters as StorybookParameters } from "@storybook/react";
+import { Decorator, StoryContext } from "@storybook/react";
 import { useMemo, useRef, useState } from "react";
 import { vscodeDark } from "@uiw/codemirror-theme-vscode";
 import { langs } from "@uiw/codemirror-extensions-langs";
@@ -6,39 +6,36 @@ import * as VibeComponents from "../../../components";
 import * as VibeComponentsNext from "../../../next";
 import * as VibeIcons from "../../../components/Icon/Icons";
 import { createPortal } from "react-dom";
-import { LiveProvider } from "react-live";
-import LivePreview from "../../components/live-preview/LivePreview";
-import styles from "./withLiveEdit.module.scss";
 import LiveEditor from "../../components/live-editor/LiveEditor";
-import { formatCode } from "./prettier-utils";
+import { formatCode } from "./utils/prettier-utils";
+import LiveContent from "./LiveContent/LiveContent";
+import { extractRenderAttributeFromCsf } from "./utils/parse-csf-utils";
 
 const globalScope = { ...VibeComponents, VibeIcons, VibeNext: VibeComponentsNext };
 
-type Parameters = StorybookParameters & {
-  liveEdit?: {
-    scope?: Record<string, unknown>;
-    isEnabled?: boolean;
-  };
-};
-
-function getInitialCodeValue(code: string, shouldPrintError: boolean): string {
+function getInitialCodeValue(source: string, shouldPrintError: boolean): string {
   try {
+    // need to wrap with parentheses to avoid syntax errors
+    const code = extractRenderAttributeFromCsf(`(${source})`);
     return formatCode(code);
   } catch (e) {
     if (shouldPrintError) {
       console.error(e);
     }
-    return code;
+    return source;
   }
 }
 
-const withLiveEdit: Decorator = (Story, { id, parameters, viewMode }: Parameters) => {
+const withLiveEdit: Decorator = (Story, context: StoryContext) => {
+  const { id, parameters, viewMode, moduleExport } = context;
   const scope = { ...globalScope, ...parameters.docs?.liveEdit?.scope };
   const canvasEditorContainer = useMemo(() => document.getElementById(id), [id]);
   const shouldAllowLiveEdit = viewMode === "docs" && parameters.docs?.liveEdit?.isEnabled && !!canvasEditorContainer;
 
-  const originalCode = useRef<string>(extractCodeFromSource(parameters.docs.source?.originalSource) || "");
-  const [code, setCode] = useState<string>(getInitialCodeValue(originalCode.current, shouldAllowLiveEdit));
+  const originalCode = useRef<string>(
+    getInitialCodeValue(parameters.docs.source?.originalSource || "", shouldAllowLiveEdit)
+  );
+  const [code, setCode] = useState<string>(originalCode.current);
   const [dirty, setDirty] = useState<boolean>(false);
 
   const handleChange = (newVal: string) => {
@@ -53,12 +50,7 @@ const withLiveEdit: Decorator = (Story, { id, parameters, viewMode }: Parameters
   return (
     <>
       {dirty ? (
-        <>
-          <div className={styles.modifiedVersionIndicator}>Modified Version</div>
-          <LiveProvider code={code} scope={scope} enableTypeScript>
-            <LivePreview />
-          </LiveProvider>
-        </>
+        <LiveContent code={code} scope={scope} decorators={moduleExport.decorators || []} context={context} />
       ) : (
         <Story />
       )}
@@ -82,19 +74,5 @@ const withLiveEdit: Decorator = (Story, { id, parameters, viewMode }: Parameters
     </>
   );
 };
-
-function extractCodeFromSource(csfSource: string): string {
-  // capture "render:" from the string
-  if (!csfSource) {
-    return "";
-  }
-  const match = csfSource.match(/render:\s*(?:\(\)\s*=>\s*)?([\s\S]*?)(?=\s*,\s*[\w]+:\s*|}$)/);
-
-  if (!match?.[1]) {
-    return "";
-  }
-
-  return match[1].trim();
-}
 
 export default withLiveEdit;